test reorganization part 1
